Move methods of Recording from RoomService to RecordingService
